asr-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6

faster-whisper==1.1.0
requests>=2.31.0
pydub==0.25.1
python-ulid==3.0.0
httpx==0.28.1
